Change % to MOD() in pseudocode
In some languages (Java is one example), % is not the modulo operator but the remainder operator. The difference being that using the remainder operator on a negative input leads to a negative result, something we want to avoid in the exact calculation that is referenced here (we want LONGITUDE to be in range 0..360, not -360..360). As the example is just pseudocode anyway, it's best to avoid this ambiguity by using the more explicit 'MOD()', to avoid faulty implementations down the line.